+++ test-infra/variables.tf
@@ -1,0 +1,15 @@
+variable "ec2_key_name" {
+  type = string
+}
+
+variable "ssh_key_path" {
+  type = string
+}
+
+variable "aws_credentials_path" {
+  type = string
+}
+
+variable "flintrock_config_path" {
+  type = string
+}