+++ tests/test_flintrock.py
@@ -1,16 +1,20 @@
+import os
+
 # External modules
 import pytest
 
 # Flintrock modules
 from flintrock.exceptions import (
-    UsageError
+    Error,
+    UsageError,
 )
 from flintrock.flintrock import (
     option_name_to_variable_name,
     variable_name_to_option_name,
     option_requires,
     mutually_exclusive,
-    get_latest_commit
+    get_latest_commit,
+    validate_download_source,
 )
 
 
@@ -129,6 +133,11 @@ def test_mutually_exclusive():
             scope=locals())
 
 
+@pytest.mark.xfail(
+    reason="This test often fails on Travis CI for unknown reasons.",
+    raises=Exception,
+    condition=(os.environ.get('TRAVIS') == 'true'),
+)
 def test_get_latest_commit():
     sha = get_latest_commit("https://github.com/apache/spark")
     assert len(sha) == 40
@@ -138,3 +147,20 @@ def test_get_latest_commit():
 
     with pytest.raises(Exception):
         get_latest_commit("https://github.com/apache/nonexistent-repo")
+
+
+@pytest.mark.xfail(
+    reason=(
+        "This test will fail whenever a new Hadoop or Spark "
+        "release is made, which is out of our control."
+    ),
+    raises=Error,
+)
+def test_validate_valid_download_source():
+    validate_download_source("https://www.apache.org/dyn/closer.lua?action=download&filename=hadoop/common/hadoop-2.8.5/hadoop-2.8.5.tar.gz")
+    validate_download_source("https://www.apache.org/dyn/closer.lua?action=download&filename=spark/spark-2.4.5/spark-2.4.5-bin-hadoop2.7.tgz")
+
+
+def test_validate_invalid_download_source():
+    with pytest.raises(Error):
+        validate_download_source("https://www.apache.org/dyn/closer.lua?action=download&filename=hadoop/common/hadoop-2.8.3/hadoop-2.8.3.tar.gz")